Avoid conflicts between tag files that are located in different directories and therefore are considered to belong to different implicit tag libraries.
The path of a tag file is now reflected in the directory in which the
corresponding tag handler is stored, and in the name of the package to
which the tag handler is assigned.

This change also avoids conflicts between tag files with the same name
and path under /META-INF/tags/ and /WEB-INF/tags/, by adding "meta" or
"web", respectively, to the directory and package names of the
respective tag handlers.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@305361 13f79535-47bb-0310-9956-ffa450edef68