Few changes to make Bootstrap useable as an MBean ( via modeler ) This allows tomcat to be embeded using its own class loader hierarchy ( i.e. the embedding app will not need to have tomcat classes in the main loader ).
Given that most of the admin operations will be exposed via MBeans
there is little need for the app to see tomcat internals. However
some interfaces or classes may be needed. This can be tuned by adding
classes and libs to the parent loader.

In general all tomcat .jars should work in either parent loader or in
the server loader.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301365 13f79535-47bb-0310-9956-ffa450edef68