Servlet 2.4 section 5.4 will be modified:
A servlet should set the locale and the character encoding of a
response. The locale is set using the ServletResponse.setLocale method,
and communicated to the client using the Content-Language header. The
character encoding can be set explicitly using the ServletResponse
methods setCharacterEncoding and setContentType, or implicitly using the
ServletResponse.setLocale method, and is communicated to the client
using the charset parameter of the Content-Type header. Explicit
specifications take precedence over implicit specifications.
[...]
The character encoding should be specified before the getWriter method
of the ServletResponse interface is called; otherwise the default
ISO-8859-1 is used.

---------------------
That means if setContentType is called, then setLocale should do reset the content type. Same for setCharacterEncoding. If getWriter is called, then ignore any call to setContentType, setCharacterEncoding and setLocale.

Please review.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@298208 13f79535-47bb-0310-9956-ffa450edef68