Remove the dependency of jasper on xerces.
The "default" if xerces is not detected is to return UTF8 for encoding.
This is probably not the best answer - the code should at least parse
the file or try to do a simpler version of what xerces is doing, but
it is far too complex.

Now at least it works without xerces.

The detection should be probably a commons package.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@305438 13f79535-47bb-0310-9956-ffa450edef68