Few changes to give more info in the jmx console.
Switched StandardManager to c-l. There are still few dozens files that
need to be converted - and we need to find a naming scheme for the logs
that includes the host/path/component.

The additions provide info about detected duplicates, total number of
sessions that were generated, number of sessions that expired, number
of sessions that failed because of maxActive.

Also operations to show the sessions and get basic info ( expire ).
I think this would simplify some debugging and be more informative.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301445 13f79535-47bb-0310-9956-ffa450edef68