Bug fixes and minor improvements to the manager and html manager web applications.  These patches fix as many of the 11 bugs I found for the manager in bugzilla that I could reproduce.
manager web application
-----------------------

Bug #5551 - Enhancement, added /serverinfo command to list Tomcat, OS, and JVM properties
Bug #7826 - Bug fix, make sure content type is set correctly for locale
Bug #8969 - Bug fix, manager remove was not removing the webapp directory/war in the appBase
Bug #13893 - Bug fix, manager remove was not removing the web application work directory

How /deploy and /undeploy work was not changed

/install war parameter will now accept a directory name or war file name without
a URL or path.  It will look for that directory or war file in the Host appBase directory.
This makes it easier to install web application directories and war files located in
the Host appBase.

/install will no longer allow installing web application directories or war files which
are not located in the Host appBase if deployXML=false. This is for security when
virtual hosting customers.

/install will now unarchive a WAR if the war is located in the Host appBase and
unpackWARs=true for the Host. To do this I had to factor out the code in
startup/HostConfig.java which expands war files out into a seperate class.

/install no longer requires a context path, if no context path is entered it
will use the name of the web application directory or war file without the
".war" extension as the context path name.

/install verifies that the context path name is the same as the web application
directory name or war file name without the ".war" extension if the directory
or war is located in the Host appbase and either autoDeploy=true or liveDeploy=true.
This change is so that the mananger works consistent with how autoDeploy and liveDeploy
work when Tomcat is restarted.

/remove will now remove the web application directory and/or war file if
the web application directory and/or war is located in the Host appBase and
either autoDeploy=true or liveDeploy=true. This change is so that the mananger
works consistent with how autoDeploy and liveDeploy work when Tomcat is restarted.

/remove will now remove the web application temporary work directory

Manager How-To docs
-------------------

Bug #5629 - Doc bug fix, documented that a reload does not reload the web.xml

There was no documentation for installing a Context configuration .xml file,
added documentation for this and reorganized the install docs.

Removed documentation for using an http URL for installing a war.
Support for this was removed a while ago.

Updated the documenation for changes to the manager.

html manager web application
----------------------------

Bug #13205 - Bug fix, HTML Manager threw NPE if XML Config field was empty
Bug #7826 - Bug fix, make sure content type is set correctly for locale

Wrote new html-manager-howto.xml documentation.

Added documentation to the web application, both the manager-howto.html
and the html-manager-howto.html docs. (With no tomcat-docs left nav menu)

Added the tomcat-docs header and footer to the application.

Added the ability to upload and install a war file.  This adds a
dependency on Jakarta commons-fileupload.  We will need to ask
commons to work on a release of commons-fileupload.

Added support for URL rewriting so that manager can be used if someone
has cookies disabled.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@288690 13f79535-47bb-0310-9956-ffa450edef68