Some info for the jmx console. It's nice to see the repositories.
It would also be nice to have a method to display info about how
 a class will be loaded ( from parent or what repo ) as well as last mod time.

I'll try to use WebppLoader as a basis for a loader to be used in modeler -
and merge it with the StandardLoader ( i.e. support both filesystem and
jndi sources ).

Having more JMX support in the loader will help track a lot of problems.
Wish I had more time...


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301450 13f79535-47bb-0310-9956-ffa450edef68