Make certain that the session still belong to us before returning it.
If the Servlet invalidates the session, and then later requests a new one, it is possible for the one we have to be valid because it is now being used by somebody else.  Thus we have to make certain that it still belongs to us before returning it.

Real Fix for bug #15894
Reported By: Christian Wicke cwicke@ics.uci.edu


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284787 13f79535-47bb-0310-9956-ffa450edef68