Remember when the actual session has been invalidated.
It is possible for the realSession to become valid again because it is being used by somebody else.  Thus we can't rely on the realSession on its own.

Also keep a copy of the sessionId so that we don't return somebody elses Id instead.

I'm going to some lengths here to avoid doing a String comparision for every access.

Hopefully the last fix for bug #15894
Reported By: Christian Wicke cwicke@ics.uci.edu


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284788 13f79535-47bb-0310-9956-ffa450edef68