- remove unused imports
- converted to commons-logging

- Implement the JSR77 "servlets" attribute ( required for WebModule )

- Use JSR77 name for the context.

The old name is still available, until someone updates the admin
to use the new name.

Note that the last 2 components of the name default to "none",
if tomcat is run inside a J2EEserver, the integration code will need
to set the name of the server and the name of the application for the
context.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301463 13f79535-47bb-0310-9956-ffa450edef68