I am not very sure this is the right fix - but from the spec it seems to be ok.
It seems some browsers ( some versions of IE on Mac ) have trouble parsing
the cookie in HTTPS mode. That happens because we send
"...path=/foo;Secure"

The spec on netscape site has all examples with a space after ;, and
aparently IE uses the space as delimiter.

I'll investigate further, this is very strange...


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301469 13f79535-47bb-0310-9956-ffa450edef68