Remove unneeded param.
Store the startup time instead of displaying it ( will be visible in the jmx console )

added missing getter.

Removed the final - duplication the code inside is too painfull.

BTW - in the previous commit, I also made StandardContext serializable. It won't
work ( a lot of other classes are not serializable ), but in time we may get to a point
where we can cache the configured state of the context and avoid parsing unless needed.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301487 13f79535-47bb-0310-9956-ffa450edef68