Few changes to improve the mbean registration ( and unregistration ).
There are 2 scenarios I want to support: the current one, where components are
 added/removed using internal APIs. And a JMX one, where components are created
by via JMX, and they insert themself in the running system ( for example the
admin interface creates a new connector Mbean and starts it ).

The big change is that the JMX handling is now in the component itself -
each component will be aware if its JMX identity and will take care of registering
and unregistering the childs. While this removes some of the flexibility
( i.e. JMX will be deep inside, instead of a listener ), I think the code
will be cleaner and we'll gain some other fexibilities.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301525 13f79535-47bb-0310-9956-ffa450edef68