Few fixes to make "ant build" work without the slow "deploy" target first. I still need to copy few files to get to the equivalent result.
Added an "embed" target for a minimal, embedable tomcat5.
Some of the components can be removed ( I'll need to document the deps and features implemented by
each component ) - right now it's 2.4 M.

Catalina.jar is almost 800k - it should be splited in "core+required components" and
few modules for the optional stuff ( like some valves, etc ).
Next is jasper-compiler, which is not required if precompiled jsps are used ( i.e. in
production servers or embeded apps ).
Digester,beanutils will also become optional if we finish serialization of the
Context ( that would greatly improve the startup time - and will allow "pre-processed" apps
that don't need to parse web.xml at startup ).

Regarding the (sample) scripts - each is a variation on JMX mbeans, the ant script
is working, I'll also try a standard MLET and a jboss-style services ( that will also
work with commons-modeler ).


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304313 13f79535-47bb-0310-9956-ffa450edef68