Few more fixes.
Now both reading and writting works ( and metadata, of course ).
It is a bit weird that we don't support anything but strings - but anything else
would be painfull in C. Later we can extend it to override the types and convert.

The only remaining issue before I cross this is adding getters and setters for
more info on the C side.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@298424 13f79535-47bb-0310-9956-ffa450edef68