New Servlet 2.4 PDF 2 features:
- remove web.xml extensibility support
- add support for SRV.12.3:

<quote>
The sub elements under web-app can be in an arbitrary order in this version of the specification. Because of the restriction of XML Schema, The multiplicity of the elements distributable, session-config, welcome-file-list, jsp-config, login-config, and locale-encoding-mapping-list was changed from optional to 0 or more. The containers must inform the developer with a descriptive error message when the deployment descriptor contains more than one element of session-config, jsp-con-fig, and login-config. The container must concatenate the items in welcome-file-list and locale-encoding-mapping-list when there are multiple occurrences. The multiple occurrence of distributable must be treated exactly in the same way as the single occurrence of distributable.
</quote>


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301574 13f79535-47bb-0310-9956-ffa450edef68