Few getters ( Remy: don't panic, I'm not changing the loader ! ).
The array will collect the loaders that are set into the class loader
and allow JMX to view the actual classpath ( if the loader is registered ).

I also added a bit more info in case something is missing.

Knowing the exact classpath of a webapp can be extremely usefull when debugging
strange bugs...


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301596 13f79535-47bb-0310-9956-ffa450edef68