Fixes.
The digester that reads the users.xml file expects a MemoryRealm ( there is
a cast ) - and a lot is duplicated, so I just extended it.
Added a lot of debug - I may remove it after it's tested a bit more.
( it is quite difficult to figure out what's broken without debug,
and this is a tricky subsystem )

The last part is just duplicating MemoryRealm.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301598 13f79535-47bb-0310-9956-ffa450edef68