Few small improvements.
If not appname is specified - take the name of the embeding container ( less
config for the user )

IMPORTANT: the thread class loader must be set to find login modules
in container ( like MemoryLoginModule ). For JAAS1.0 - that doesn't
help ( LoginModules must be in CLASSPATH - jboss actually has a hack
around it, but it's too complex ), but in JDK1.4 it does work fine.

If no principal is found ( but login succeded ) - use the user name to create
 a principal.

Implement the "Roles" Group - I'll revert if anyone -1.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301599 13f79535-47bb-0310-9956-ffa450edef68