Hopefully the last commit on this topic :-)
Now both deploy and build-all work.

I renamed deploy to deploy-old and used the new "build" model -
I think it is much cleaner and will allow us to simplify the
build a lot when we switch to ant1.6. In addition it may
allow us to do small changes to the layout ( like the "profiles"
we discussed earlier )

Not to mention it is faster.

You should use "build" target if you just want to build - the
default target will build webapps ( very slow ) and a lot of other
stuff.

If anyone has a problem - you can revert or use deploy-old.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304340 13f79535-47bb-0310-9956-ffa450edef68