Get the server, service self-register.
Add "serviceNames" attribute in server - this should be used instead of casting and
calling the methods directly.

In order for /admin to be used for clusters it is essential to _stop_ using any
internal tomcat class in the admin. Just do a grep o.a.catalina in admin -
it shouldn't find any import. Only JMX should be used !

The names should be stable.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301636 13f79535-47bb-0310-9956-ffa450edef68