Registration for loggers.
We do need to make a final decision about this. I didn't get any feedback on the
logger names used in context  ( with the context path as prefix of the logger name ).
IMO the logging in tomcat should be delegated via commons-logging to an external
implementation - either jdk1.4 logging or log4j.

We do need to decide on the names of the loggers and provide some default
configs, and eventually start bundling log4j or config files for jdk1.4 ( or
both ). If we do that, LoggerBase and all other classes will remain only
for backward compatibility.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301648 13f79535-47bb-0310-9956-ffa450edef68