Fix valve registration. Few methods to simplify registration.
You'll see a lot of valves ( about 130 with the default webapps ). We could filter
out "trivial" valves - but I would rather leave them in as an incentive to
eliminate some ( and reduce the stack trace size, memory use, etc :-)

For each servlet we have a pipeline with a single valve - the standard wrapper valve.
It should be possible to configure per-servlet valves  ( not sure how to do it with
server.xml, but with jmx it's not hard ) - but that doesn't mean we need all this
indirections.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301658 13f79535-47bb-0310-9956-ffa450edef68