Few small enhancements.
Make Embedded extend StandardService. It includes an Engine and connectors -
just like Service. This will allow us to harmonize the 2 use cases.
( I'm also adding code to create a StandardServer if none is found - so exactly
the same object model and JMX naming will be used in both embeded and standalone )

Also removed the test main().


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301670 13f79535-47bb-0310-9956-ffa450edef68