Few changes - hopefully no backward incompatibilities.
The most important is the use of the "threadData" property in
ThreadWithAttribute.  Previously - we created a simple thread, and
used Runnable - however this mechanism is more powerfull and allows
us to share info that was usually private.

In particular we need the thread data exposed so we can use the
ObjectName and unregister the connector and components.

I also added support for Runnable - it's a long time since I wanted to
do this. The reason we used ThreadPoolRunnable was the lack of ThreadData
in JDK1.1. However ThreadWithAttributes provide the same functionality
and is much cleaner. Now it would be possible to run more stuff in
thread pools.

Finally - I moved the code that gets info from ThreadPoolMX - there is no
need and no dependency on JMX, the registration happens in the listener.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@298537 13f79535-47bb-0310-9956-ffa450edef68