few small fixes in unregistration.
Added the "controller" - this will be the name of the component that
registers, or null if it's an outside entity.

For example, if a Container/Valve/etc is created automatically by tomcat,
then the code that creates it will set the field. When the controller stops,
it must also unregister all the components it registered ( and not more :-).

If the same Container is created by modeler from a config file or by
an embeding application by direct calls or other means - then the controller field
will not be set, which means that the container is responsible for its lifecycle.

Short version:
1. who register is responsible to unregister
2. you shouldn't unregister what you didn't registered

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301715 13f79535-47bb-0310-9956-ffa450edef68