A bit of a hack for now. When the engine is destroyed it should clean up the data that it loaded in registry - otherwise a start would see the old data.
The real solution is to stop using the static methods in Registry. The registry
is just a helper associated with the app. This will also be more secure.

In any case - for next release of modeler the static methods will be deprecated
and the method that loads explicitely an mbean-descriptor will also be deprecated.
It is much cleaner to load the descriptors on demand, from the same package.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@301722 13f79535-47bb-0310-9956-ffa450edef68