The ever-changing experimental embed startup.
Play with the realm mbeans - it works for me, let me know if it
doesn't for you.

Added a target for the commons-modeler proxy - it can now remotely manage
both tomcat and mod_jk, and aparently multiple instances as well
(obviously the domain must be different )
There are various types that are not supported - but the basic works.
Cluster management is getting closer :-)



PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304371 13f79535-47bb-0310-9956-ffa450edef68