To avoid confusions, remove the "run" target ( which was broken for some time and uses server.xml ) and rename run2.
This won't work without reverting the Engine name ( or fixing the registration )

In embed mode, we rely a lot on the domain name - a component can
locate it's parent by using the component name ( and the domain )
to construct the parent name. If Engine is identified by both domain
and name - that becomes impossible. In addition it makes it difficult
to work with multiple engines.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304379 13f79535-47bb-0310-9956-ffa450edef68