Changing to do lazy evaluation of the Servlet-Spec's SSL-attributes.
With Jean-Francois' changes to the RequestDispacher, this should work now.  Of course, it is easy enough to back-out if I'm wrong ;-).  I'm nearly 100% certain that this should work very well with JK-Coyote.  Committing to the 5.0 branch only so that I can get more eyes to see potential problems with the Standalone HTTP connector.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301882 13f79535-47bb-0310-9956-ffa450edef68