Initial version of the proposal to pass-through all attributes to the protocolHandler, and let it deal.
Now, only explictly configured attributes are passed through to the Protocol.  It is the Protocol's responsibility to have the correct default attributes configured.  This way, the HTTP/1.1 connector can default the value of "connectionTimeout" to one value, and the Jk-Coyote connector can default it to another.

For now, I'm leaving the getters and setters for all of the different attributes, so that they can still be seen from the JMX consol.  In the transistion, some of the values shown may be wrong because of the paragraph above.

For now, I'm leaving in the SSL SocketFactory support (for backwards compatibility).  It should be deprecated.  You can configure SSL support without it by e.g. specifying "property[keystore]=/path/to/my/.keystore" on the connector.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301893 13f79535-47bb-0310-9956-ffa450edef68