Fix bugtraq 4880590 Container incorrectly processes invalid URL patterns in jsp-property-groups.
This extends to non-jsp-property-group url-patterns as well as the catalina code
that handles the jsp url pattern mappings delegates to the code that handles the
servlet pattern mapping.

Consider the following url-pattern:  /somepath/*.jsp

This is invalid but accepted and works within the server.

The Servlet specification states that a valid URL pattern for an extension
mapping
like the example above must be '*.jsp' with no additional path information.

Patch Submitted by: Ryan Lubke at sun.com


git-svn-id: https://svn.apache.org/repos/asf/tomcat@301968 13f79535-47bb-0310-9956-ffa450edef68