Added support for X-Powered-By response header, as defined by the Servlet 2.4 and JSP 2.0 specs.
The Servlet 2.4 spec defines a special

  X-Powered-By: Servlet/2.4

header, which containers may add to servlet-generated responses. This
is complemented by the JSP 2.0 spec, which defines a

  X-Powered-By: JSP/2.0

header to be added (on an optional basis) to responses utilizing JSP
technology.

The goal of these headers is to aid in gathering statistical data
about the use of Servlet and JSP technology


git-svn-id: https://svn.apache.org/repos/asf/tomcat@302035 13f79535-47bb-0310-9956-ffa450edef68