When using the embedded interface (or jmx directly), context are never removed because of this condition (mapper.removeContext is never called).
Then if you re-deploy the same app, The Mapper will maps the http call to the first Mapper's context object, which is an invalid (orphan) object. The client always receives a 503 (since the context is invalid and marked as unavailable).

Removing the condition doesn't have any side effect (but fix the problem).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@302041 13f79535-47bb-0310-9956-ffa450edef68