Changing the Realm API to conform with the new requirements in the 2.4 spec.
This implementation is sub-optimal, but committing it now so that at least the 5.0.5 release will be spec-compliant.  I plan to revisit this (as well as other Authenticator optimizations).

In implementing this, I discovered that I needed to do a two-pass over the constraints (since I can't usefully do a 401 before a SSL-redirect).  That's the reason for using SecurityConstraint [].


git-svn-id: https://svn.apache.org/repos/asf/tomcat@302052 13f79535-47bb-0310-9956-ffa450edef68