Simplify the way the URLClassLoader gets initialized.
In the case where jspc is launched from the command line (instead of
ant), the same class path JARs were added 3 (!) times to the
URLClassLoader's URL array.

Notice that the URLClassLoader returned by
this.getClass().getClassLoader() is already initialized with all the
URLs from java.class.path, so those URLs don't need to be added
to the classpath.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@305763 13f79535-47bb-0310-9956-ffa450edef68