Fixing getters for JMX.
It looks like a big commit, but it really shouldn't be very much in terms of functionality.  All default values are (almost) removed from the Connector.  This may break other components that were depending on them.  However that is why I'm committing now (so I can get more eyes to find the broken components).

I've kludged getScheme for now, so that Tomcat is more or less functional.  However, it looks like there is still a fair amount of cleanup to do in getting the default values set for the protocols.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@302163 13f79535-47bb-0310-9956-ffa450edef68