Fixed Bugtraq 4863026 ("JSP error page mechanism fails intermittently to display contents of error page")
Problem was caused by the fact that one of the tests in the suite
(JspFactoryTest.jsp) released a custom page context (of type
com.sun.ts.tests.jsp.common.util.SimpleContext), like this:

  SimpleContext context = new SimpleContext();
  factory.releasePageContext(context);

The implementation of
org.apache.jasper.runtime.JspFactoryImpl.releasePageContext would add
the page context back to the pool of page contexts. Page contexts
would be retrieved from the pool as follows:

  pc = (PageContextImpl) pool.get()

Clearly, if the page context returned was an instance of
com.sun.ts.tests.jsp.common.util.SimpleContext, the above assignment
would result in a ClassCastException.

The fix is to return only those page contexts that are instances of
org.apache.jasper.runtime.PageContextImpl to the pool, since we don't
have any control over the implementation of custom page contexts


git-svn-id: https://svn.apache.org/repos/asf/tomcat@305815 13f79535-47bb-0310-9956-ffa450edef68