Fix potential problem with dying threads when the pool is empty.
Functionally, this isn't that much different from what Remy posted on Bug #21763.  However (at least on high-end machines) it should be able to respond faster to a released CM.  Since the case that all threads are dead may be able to be handled, be less aggressive on checking this case.

Also adding debug statements, since I'm not convinced that this is the fix for 21763 (so I'm going to do some more testing before I close the bug).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@298876 13f79535-47bb-0310-9956-ffa450edef68