The DefaultContext docs listed support for a Listener but the code to suppor this was not implemented.
Added support for nesting a Context Listener within a
DefaultContext.

Added support for nesting a Webapp Loader within a
DefaultContext.

The importDefaultContext() had to be split into two
methods.  The installDefaultContext() method was added.
This is due to sequencing issues when instantiating
a Context.  Some things needed to occur earlier than
when the importDefaultContext() is called. And some
things needed to occur during the normal
importDefaultContext().

No interfaces in org.apache.catalina were changed,
though they could be in Tomcat 5 since there has
not been a final release.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@288869 13f79535-47bb-0310-9956-ffa450edef68