Fix JkCoyote to properly handle the distinction between localName/Port and serverName/Port.
As always with mod_jk, you can lose strict Servlet-Spec compliance by playing with your Apache settings.  However if ConnonicalName=on or DNS, then this is correct.  If it's not, then you obviously don't want this to be correct ;-).  It also seems to be the value that IIS is sending.

Thanks to Remy for his great parseHost routine (which I had to modify slightly, because of design differences in JkCoyote and HTTP11Coyote).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/connectors/trunk@299061 13f79535-47bb-0310-9956-ffa450edef68