First attempt to get Tomcat compliant with the Servlet 2.4 Spec for AA.
This one isn't very efficient (but not much worse than it was before :).  Longer term, I'm looking to modify the j-t-c mapper to be able to return a Collection of mapped objects, and use that.  In that case it could also be used for Filter mappings.

This version is just to get a Spec compliant version out there that can be tested.

Reported By: Philippe Leothaud <philippe.leothaud@wanadoo.fr>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@302496 13f79535-47bb-0310-9956-ffa450edef68