Getting back to strict compliance with the Spec for AAA.
My last patch included and illegal optimization (check methods before checking path).  Now methods are correctly checked after the path has been checked.

Also including some of Remy's optimizations from the Filter mapping code.

This doesn't include Phillepe's idea of caching the mappings.  This is largely because the submitted patch is broken (it assumes that the Realm is configured at the Context level), but I like the idea.  I'd like to add a config switch to cache security-constraint mapping (defaulting to what the list's opinions are).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@302501 13f79535-47bb-0310-9956-ffa450edef68