Exempt welcome pages from any security-constraint checks.
The Servlet 2.4 spec does not require this (and there are no CTS tests
for this), but it seems like a reasonable enhancement. I was told that
the upcoming maintenance release of the Servlet spec is going to
clarify this.

If this change is controversial, I'll back it out for the time being,
until it is backed by the Servlet spec. Please let me know.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@302844 13f79535-47bb-0310-9956-ffa450edef68