Reset the Context CL correctly after an include.  Also, adding a PostRequest so the the Context CL is in a defined state after the request finishes.
The rule is that if the Child has a Parent, then the Parent is the request that included.  Otherwise, the including request is the top-level one.

There are still CL issues with cross-context RD reqests, but this at least fixes the bug that was there with same-context RD requests.

Fix for Bug #28590
Reported By: Hugh J. Lee hughjlee@yahoo.com


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284912 13f79535-47bb-0310-9956-ffa450edef68