Clean up the API a bit to make it easier to improve.
The big change is that the 'source' attribute of MsgContext is now a JkChannel instead of an arbitrary JkHandler.  This allows the Actions to bypass the 'invoke' path (which is really for incoming requests), and makes Request registration easier.

Other than Request registration, there really isn't much functionality change here.  However, now JK finally plays nice with the Manager status servlet (at least with the Socket and Unix channels).  Also, the memory leak with the JNI channel is finally gone (since it doesn't implement Request registration).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@299444 13f79535-47bb-0310-9956-ffa450edef68