Reverted fix for Bugzilla 29763 ("The encoding of jsp document").
We're supposed to throw an error only if there is an encoding declaration in the XML prolog that does not match the page encoding from the page directive or jsp-config. We must not throw any error if there is no XML prolog or encoding declaration in the prolog.

The motivation for this is to make sure that JSP 1.2 pages with a page directive specifying a page encoding other than the default will continue to work in JSP 2.0 in most cases.

The code is currently going through lots of pains to determine whether the encoding was specified in the XML prolog, or autodetected, to decide whether an error should be thrown.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@306015 13f79535-47bb-0310-9956-ffa450edef68