Undid previous change, as in the case where a servlet has a jsp-file and also declares some init params, as in:
  <servlet>
    <servlet-name>xxx</servlet-name>
    <jsp-file>/xxx.jsp</jsp-file>
    <init-param>
      <param-name>name1</param-name>
      <param-value>value1</param-value>
    </init-param>
  </servlet>

it needs its *own* JspServlet instance that it can initialize with its
own params. Sharing of JspServlet instance is not possible in this case.

Will have to come up with a better solution against loss of monitoring
info (the JspServlet that handles the above jsp-file currently is not
registered with JMX).


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@303358 13f79535-47bb-0310-9956-ffa450edef68