Expose any errors on an included resource.
For example, a JSP with this include action:
  <jsp:include page="nonexistent">
or
  <jsp:include page="nonexistent.jsp">
where nonexistent[.jsp] does not exist, currently returns silently, hiding the
fact that the resource to be included does not exist.

This patch returns a 404 with the name of the nonexistent resource.

Yes, SRV.8.3 ("The Include Method") mentions that

  "it [the target servlet] cannot set headers or call any method that
  affects the headers of the response. Any attempt to do so must be
  ignored."

but i don't think it is referring to the error case.

Let me know if you see any problems.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@303385 13f79535-47bb-0310-9956-ffa450edef68