Clarified description of 'reloading' init param.
I'm about to remove this param altogether, as it is redundant and
misleading (it's only evaluated in deployment mode):

  reloading=true should be covered by checkInterval>0
  reloading=false should be covered by checkInterval<=0

Let me know if there are any objections to removing this param and
associated APIs. This is currently only used in JspRuntimeContext to
determine whether thread for background compiles should be spawn, and may
be replaced as described above.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@303394 13f79535-47bb-0310-9956-ffa450edef68