Add an asynchronous byte buffer based access log valve. The valve use a direct byte buffer to store and save the log. I didn't use the backgroundProcess since it's affect all components, but this should be revisited.
Benchmarking this valve using jservlet (scalabilty) and trade2 (throughput) demonstrated a 13% improvement over the current access log valves. Increasing the writerInterval improve performance too.

Note: I've tried to use a direct CharBuffer instead of a StringBuffer, and its slower that the current approach (and the log contains garbage, thank to a bug in NIO).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@303504 13f79535-47bb-0310-9956-ffa450edef68