Fix for 32832 ("request.getSession(false) fails to return null").
I believe this bug has been valid: If the session in the foreign
context has been invalidated, it must not be returned. A comment in
the code actually stated that the current session be returned "if it
exists and is valid", but the isValid() check on the session was
missing, and is being added by this commit.
Also, a session is now created in the foreign context only if 'create' is TRUE.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@303631 13f79535-47bb-0310-9956-ffa450edef68