Consider the case where original request was mapped to welcome page. In this case, the mapped welcome page (and not the original request URI!) needs to be the target of hasResourcePermission().
This is consistent with the change that had been made in findSecurityConstraints().

BTW, shouldn't request.getDecodedRequestURI() return the mapped
welcome page (instead of the original URI) in this case?
In other words, shouldn't the path passed to
  mappingData.requestPath.setString(pathStr)
in Mapper.java be propagated to the request object associatd with the
mappingData?


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@303729 13f79535-47bb-0310-9956-ffa450edef68