No longer ignore the wrapper class passed to Context.setWrapperClass().
I think the original intent of this method was to pass a class that
implements the Wrapper interface. However, there are too many
dependencies in the code right now that assume the children of
StandardContext to be instances of StandardWrapper (instead of
Wrapper), which means the class passed to StandardContext.setWrapperClass()
really must be a subclass of StandardWrapper.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@303874 13f79535-47bb-0310-9956-ffa450edef68