Copy the idea from the APR Connector, and delay reading the initial request body packet until the Servlet asks for it.
This reduces the chance that the socket read will block uselessly waiting for the data to be available.  You can restore the previous behavior by setting request.delayInitialRead="false" on the <Connector>.

Of course, if Bill R. decides to port his C-L patch for proxy_http to mod_jk, the value of this setting won't matter :).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@300984 13f79535-47bb-0310-9956-ffa450edef68