Well, NIO ByteBuffer is the most horrible interface in java. Whoever invented the flip() interface - and a buffer where you can't know where it starts and where it ends and where is the data - should be ashamed...
The only way we can use the buffer is by keeping track of start and end
as before, anything that uses position() or limit() will never know if
it's reading or writting or if the buffer has flipped. 

The use of ByteBuffer is still a good thing - if you ignore the stupid
interface, as is the only way to work with direct and mapped byte[], and
maybe in a smarter VM get some buffers that are not relocated. It is
also usefull because it can work with the usefull parts of nio - like
charset and channels.




git-svn-id: https://svn.apache.org/repos/asf/tomcat@330199 13f79535-47bb-0310-9956-ffa450edef68