Add a project here as well, for checkout to work.
The only solution I found so far to use eclipse and svn is to do a
checkout from the repostiory - 'share project' or import existing
project - or any other thing - fails. This seems to kind-of-work for me.
BTW - I can't even see the current/ 

How is everyone using eclipse with svn ??


git-svn-id: https://svn.apache.org/repos/asf/tomcat@332738 13f79535-47bb-0310-9956-ffa450edef68