New endpoint - it's a long time since I had this planned, now I got the chance.
I think this is the _RIGHT_ way to do the pool and threading - not run any benchmark, but I have 
a feeling it will be good ( in particular on MP machines !). 

Each thread is accepting - the kernel decides who gets to process the request ( may accept at 
the same time on multiple CPUs AFAIK ) - no thread switching or lock or anything.

I'll use this as a base for the NIO stuff


git-svn-id: https://svn.apache.org/repos/asf/tomcat@348659 13f79535-47bb-0310-9956-ffa450edef68