Fix handling of the special first Body-Chunk packet.
Somebody that knows the APR interface better than me could probably clean this up (e.g. swallow an unread first Body-Chunk in recycle instead of in process).  However, it probably doesn't actually make much difference in terms of processing speed.

I haven't actually tested this, but now it's doing much the same thing as the JK/Java Connector, so it should be fine.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@360466 13f79535-47bb-0310-9956-ffa450edef68