Add new ALV attribute "renameOnRotate" (Default:false).
If set to "true", the "fileDateFormat" will not be part
of the current log file. Only at the time of log rotation
the file is renamed to the final name including the
"fileDateFormat".

This mimics the behavior e.g. of Log4J and similar frameworks,
where the active file does not have the timestamp in the file
name.

Pro: current file has stable name, all files with
     timestamp in name are old.
Con: Slightly more complex impl.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1414889 13f79535-47bb-0310-9956-ffa450edef68